copyright added & sample app moved into new dir
