+++ pjsip-apps/src/pjsua2/ios-swift-pjsua2/ios-swift-pjsua2.xcodeproj/project.xcworkspace/contents.xcworkspacedata
