+++ pjsip-apps/src/pjsua2/ios-swift-pjsua2/ios-swift-pjsua2.xcodeproj/xcshareddata/xcschemes/ios-swift-pjsua2.xcscheme
