+++ pjsip-apps/src/pjsua2/ios-swift-pjsua2/ios-swift-pjsua2/Assets.xcassets/AccentColor.colorset/Contents.json
