+++ pjsip-apps/src/pjsua2/ios-swift-pjsua2/ios-swift-pjsua2/Assets.xcassets/AppIcon.appiconset/Contents.json
