+++ pjsip-apps/src/pjsua2/ios-swift-pjsua2/ios-swift-pjsua2/Controller/IncomingViewController.swift
@@ -1,0 +1,52 @@
+/*
+ * Copyright (C) 2012-2012 Teluu Inc. (http://www.teluu.com)
+ * Contributed by Regis Montoya (aka r3gis - www.r3gis.fr)
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ */
+
+import UIKit
+
+
+
+class IncomingViewController: UIViewController {
+
+    var incomingCallId : String = ""
+    @IBOutlet weak var callTitle: UILabel!
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        title = "Gelen Arama"
+        callTitle.text = incomingCallId
+        
+        CPPWrapper().call_listener_wrapper(call_status_listener_swift)
+    }
+    
+    override func viewDidDisappear(_ animated: Bool) {
+        CPPWrapper().hangupCall();
+    }
+    
+    
+    @IBAction func hangupClick(_ sender: UIButton) {
+        CPPWrapper().hangupCall();
+        self.dismiss(animated: true, completion: nil)
+    }
+    
+    
+    @IBAction func answerClick(_ sender: UIButton) {
+        CPPWrapper().answerCall();
+    }
+
+}
