+++ pjsip-apps/src/pjsua2/ios-swift-pjsua2/ios-swift-pjsua2/PJSua2/CustomPJSUA2.hpp
@@ -1,9 +1,22 @@
-//
-//  pjsua2_emre.hpp
-//  ios-swift-pjsua2
-//
-//  Created by Emre on 7.01.2021.
-//
+/*
+ * Copyright (C) 2012-2012 Teluu Inc. (http://www.teluu.com)
+ * Contributed by Regis Montoya (aka r3gis - www.r3gis.fr)
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 2 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
+ */
+
 #include <string>
 #include <pjsua2.hpp>
 #include <dispatch/dispatch.h>
