--- pjsip-apps/src/samples/ios-swift-pjsua2/ios-swift-pjsua2/ios-swift-pjsua2-Bridging-Header.h
@@ -1,5 +1,0 @@
-//
-//  Use this file to import your target's public headers that you would like to expose to Swift.
-//
-
-#import "wrapper.h"
